Move multiclass wrappers into a single source file
