+++ python/cuml/multiclass/__init__.py
@@ -13,5 +13,5 @@
 # limitations under the License.
 #
 
-from cuml.multiclass.one_vs_one import OneVsOneClassifier
-from cuml.multiclass.one_vs_rest import OneVsRestClassifier
+from cuml.multiclass.multiclass import OneVsOneClassifier
+from cuml.multiclass.multiclass import OneVsRestClassifier
