+++ python/cuml/multiclass/multiclass.py
@@ -48,3 +48,36 @@ class OneVsOneClassifier(sklearn.multiclass.OneVsOneClassifier):
         with cuml.using_output_type('numpy'):
             df = super(OneVsOneClassifier, self).decision_function(X)
         return df
+
+
+class OneVsRestClassifier(sklearn.multiclass.OneVsRestClassifier):
+    """ Wrapper around Sckit-learn's class with the same name. This wrapper
+    accepts any array type supported by cuML and converts them to numpy if
+    needed to call the corresponding sklearn routine.
+
+    See issue https://github.com/rapidsai/cuml/issues/2876 for more info about
+    using Sklearn meta estimators.
+    """
+    def __init__(self, estimator, *args, n_jobs=None):
+        super(OneVsRestClassifier, self).__init__(estimator, *args,
+                                                  n_jobs=n_jobs)
+
+    def fit(self, X, y):
+        X, _, _, _, _ = input_to_host_array(X)
+        y, _, _, _, _ = input_to_host_array(y)
+        with cuml.using_output_type('numpy'):
+            return super(OneVsRestClassifier, self).fit(X, y)
+
+    def predict(self, X):
+        # out_type = self.estimator._get_output_type(X)
+        X, _, _, _, _ = input_to_host_array(X)
+        with cuml.using_output_type('numpy'):
+            preds = super(OneVsRestClassifier, self).predict(X)
+        return preds
+
+    def decision_function(self, X):
+        # out_type = self.estimator._get_output_type(X)
+        X, _, _, _, _ = input_to_host_array(X)
+        with cuml.using_output_type('numpy'):
+            df = super(OneVsRestClassifier, self).decision_function(X)
+        return df
